Refactoring detectors inside submodules
